Αλλαγές στα create των πινακων και στο workspace.html
Έγιναν αλλαγές για τα σωστά create των πινάκων. Επίσης βασικές αλλαγές στο workspace για να είναι λειτουργικό

+++ PythonProject/Ptyxiaki/loadsource.py
@@ -54,7 +54,7 @@ def create_loadsource_table(cursor, db):
 
         cursor.execute("""
             CREATE TABLE loadsource (
-                LID INT NOT NULL,
+                LID tinyint NOT NULL,
                 name_loadsource VARCHAR(50) NOT NULL,
                 PRIMARY KEY (LID),
                 UNIQUE (name_loadsource)
